Script to find non-existant or zero-byte recordings
This will be run from a mythsystem hook of REC_STARTED.  If the file doesn't
exist in 30s, it belches.  If the file isn't > 0 bytes within 10s of existance,
it belches.

To be added: add a "delete recording and allow re-record" message, and email
me for notification.  Later on, one the hardware's ready, also power cycle
the HD-PVR if it's the failing source.